+++ src/resource/compile_options.h
@@ -159,6 +159,9 @@ struct CompileOptions
 
 	///
 	void read_output(StringStream &ss, Process &pr);
+
+	/// Returns the target platfom as a string.
+	const char *platform_name();
 };
 
 } // namespace crown